SDK-215 transfer getting secrets from JSON to SecretsDataGenerator
