+++ src/main/java/com/incountry/residence/sdk/tools/keyaccessor/key/SecretsDataGenerator.java
@@ -1,6 +1,7 @@
 package com.incountry.residence.sdk.tools.keyaccessor.key;
 
-import com.incountry.residence.sdk.tools.JsonUtils;
+import com.google.gson.Gson;
+import com.google.gson.JsonSyntaxException;
 import com.incountry.residence.sdk.tools.exceptions.StorageClientException;
 
 import java.nio.charset.StandardCharsets;
@@ -13,6 +14,7 @@ import java.util.List;
 public class SecretsDataGenerator {
 
     public static final int DEFAULT_VERSION = 0;
+    private static final String MSG_ERR_INCORRECT_SECRETS = "Incorrect JSON with SecretsData";
 
     private SecretsDataGenerator() {
     }
@@ -39,6 +41,31 @@ public class SecretsDataGenerator {
      * @throws StorageClientException when parameter validation fails
      */
     public static SecretsData fromJson(String secretsDataJson) throws StorageClientException {
-        return JsonUtils.getSecretsDataFromJson(secretsDataJson);
+        SecretsData result;
+        try {
+            SecretsDataContainer container = new Gson().fromJson(secretsDataJson, SecretsDataContainer.class);
+            List<SecretKey> secrets = new ArrayList<>();
+            if (container.secrets != null) {
+                for (SecretKeyContainer key : container.secrets) {
+                    secrets.add(new SecretKey(key.secret.getBytes(StandardCharsets.UTF_8), key.version, key.isKey, key.isForCustomEncryption));
+                }
+            }
+            result = new SecretsData(secrets, container.currentVersion);
+        } catch (JsonSyntaxException | NullPointerException e) {
+            throw new StorageClientException(MSG_ERR_INCORRECT_SECRETS, e);
+        }
+        return result;
+    }
+
+    private static class SecretsDataContainer {
+        List<SecretKeyContainer> secrets;
+        Integer currentVersion;
+    }
+
+    private static class SecretKeyContainer {
+        String secret;
+        Integer version;
+        boolean isKey;
+        boolean isForCustomEncryption;
     }
 }
